backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
PyPDF2==3.0.1
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.3.7
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.0
reportlab==4.0.7
markdown==3.5.1
